Downcast instead of returning tip with an error
Trying to have a unified way of printing errors different
types (CLI's and the library's) along with their respective tip messages
to the user led to convoluted code.

Simply downcasting the error to its original type at the time of
printing errors reduces the complexity drastically, and is certainly a
preferred way of handling errors. As this allows access to the Tip trait
on either BlibError or BlightError, whichever is successfully
downcasted, which can then be used to print the tip message if it is present.