Bump ver to 0.4.1
Note: the changes leading up to this minor version bump include what
might be considered as breaking changes. However, it's the Tip trait
whose interface has been changed. This trait, although exposed in the
public API is not meant to be used by the user. It solely exists for
the CLI, therefore, the version bump will only be a minor one.