Turn helper fn sweep into a method on Device
The helper function 'sweep' is now called 'sweep_write', and is available
as a method on a Device struct instance.

This greatly simplifies the function signature, and makes more sense
when it is placed next to the Device::write_value method instead of
leaving it as an independent helper function.

Direction is internally determined to further simplify the fn signature,
making it essentially the same as Device::write_value.