Rewind and write to opened file in sweep_write
Older implementation relied on Device::write_value to write updated
brightness value on each iteration of the loop, which in-turn uses
std::fs::write. It is more efficient to open the brightness file once in
write mode and write to it on each iteration instead.

This may reduce any visible lag in brightness changes, leading to a
smoother sweep effect.